+++ backend/chainlit/auth.py
@@ -5,7 +5,7 @@ from typing import Any, Dict
 import jwt
 from chainlit.config import config
 from chainlit.data import get_data_layer
-from chainlit.oauth_providers.providers import get_configured_oauth_providers
+from chainlit.oauth.providers import get_configured_oauth_providers
 from chainlit.user import User
 from fastapi import Depends, HTTPException
 from fastapi.security import OAuth2PasswordBearer
