+++ backend/chainlit/config.py
@@ -19,7 +19,7 @@ from typing import (
 
 import tomli
 from chainlit.logger import logger
-from chainlit.oauth_providers.oauth_provider import OAuthProvider
+from chainlit.oauth.oauth_provider import OAuthProvider
 from chainlit.translations import lint_translation_json
 from chainlit.version import __version__
 from dataclasses_json import DataClassJsonMixin
@@ -295,7 +295,7 @@ class CodeSettings:
     ] = None
     # Callbacks for authenticate mechanism
     custom_authenticate_user: Optional[Callable[[str], Awaitable["User"]]] = None
-    custom_oauth_provider: Optional[Type[OAuthProvider]] = None
+    custom_oauth_provider: Optional[Callable[[], Type[OAuthProvider]]] = None
     on_logout: Optional[Callable[["Request", "Response"], Any]] = None
     on_stop: Optional[Callable[[], Any]] = None
     on_chat_start: Optional[Callable[[], Any]] = None
