+++ backend/chainlit/oauth/azure_ad_hubrid_oauth_provider.py
@@ -2,7 +2,7 @@ import base64
 import os
 
 import httpx
-from chainlit.oauth_providers.oauth_provider import OAuthProvider
+from chainlit.oauth.oauth_provider import OAuthProvider
 from chainlit.secret import random_secret
 from chainlit.user import User
 from fastapi import HTTPException
