+++ backend/chainlit/oauth/github.py
@@ -2,7 +2,7 @@ import os
 import urllib.parse
 
 import httpx
-from chainlit.oauth_providers.oauth_provider import OAuthProvider
+from chainlit.oauth.oauth_provider import OAuthProvider
 from chainlit.user import User
 from fastapi import HTTPException
